refactor(gateway): split state view modules
